+++ src/utils/geometry_utils.cpp
@@ -10,7 +10,7 @@ namespace Planners
             float calculatePathLength(const CoordinateList &_path, const double &_resolution)
             {
                 float len = 0;
-                for (int i = 1; i < _path.size(); i++)
+                for (long unsigned int i = 1; i < _path.size(); i++)
                 {
                     len += sqrtf(pow((_path[i].x - _path[i - 1].x) * _resolution, 2) +
                                  pow((_path[i].y - _path[i - 1].y) * _resolution, 2) +
